refactor: pr comments - optimization
