+++ src/components/organisms/introMainBlock.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable @liferay/empty-line-between-elements */
 import React, { useState } from "react";
 
 import { useTranslation } from "react-i18next";
