+++ src/components/organisms/topbar/project/navigation.tsx
@@ -3,7 +3,7 @@ import React, { useCallback, useEffect, useMemo } from "react";
 import { useLocation, useNavigate, useParams } from "react-router-dom";
 
 import { mainNavigationItems } from "@src/constants";
-import { useCacheStore } from "@src/store/useCacheStore";
+import { useCacheStore } from "@src/store";
 import { cn } from "@src/utilities";
 
 import { Button, IconSvg } from "@components/atoms";
