+++ src/components/pages/intro.tsx
@@ -23,7 +23,7 @@ export const Intro = () => {
 				<CatDashboardImage className="absolute -bottom-6 -right-5 hidden minHeightLg:block" />
 			</div>
 
-			{/* eslint-disable-next-line tailwindcss/no-custom-classname, @liferay/empty-line-between-elements */}
+			{/* eslint-disable-next-line tailwindcss/no-custom-classname */}
 			<div className="resize-handle-horizontal z-10 -ml-2 w-1 cursor-ew-resize transition hover:bg-gray-750" />
 
 			<div style={{ width: `${100 - (leftSideWidth as number)}%` }}>
