+++ src/locales/en/errors/translation.json
@@ -74,7 +74,7 @@
 	"projectBuildFailed": "Project build failed",
 	"projectDeployFailed": "Project deploy failed",
 	"activityLogsFetchError": "Couldn't fetch the activity logs",
-	"activityLogsFetchErrorExtended": "Couldn't fetch the activity logs",
+	"activityLogsFetchErrorExtended": "Couldn't fetch the activity logs, {{error}}",
 	"outputLogsFetchError": "Couldn't fetch the output logs",
-	"outputogsFetchErrorExtended": "Couldn't fetch the output logs"
+	"outputLogsFetchErrorExtended": "Couldn't fetch the output logs, {{error}}"
 }
