+++ src/store/index.ts
@@ -8,5 +8,6 @@ export { useConnectionCheckerStore } from "@store/useConnectionCheckerStore";
 export { useDrawerStore } from "@store/useDrawerStore";
 export { useManualRunStore } from "@store/useManualRunStore";
 export { useProjectValidationStore } from "@store/useProjectValidationStore";
-export { useActivitiesCacheStore } from "@store/useActivitiesCacheStore";
-export { useOutputsCacheStore } from "@store/useOutputsCacheStore";
+export { useActivitiesCacheStore } from "@src/store/cache/useActivitiesCacheStore";
+export { useOutputsCacheStore } from "@src/store/cache/useOutputsCacheStore";
+export { useCacheStore } from "@src/store/cache/useCacheStore";
